Add support to search after (#198)
This commit introduces the `search_api` option to enable support for both search_after and scroll.

`auto`: Default value. It uses `search_after` for Elasticsearch 8 and newer versions, otherwise it falls back to the `scroll` API
`search_after`: Create a point in time for search, which is the recommended way to do paginated search since 7.10
`scroll`: Uses scroll API to search, which is the previous way to search

When search_after is utilized, if the query doesn't specify a sort field, a default sort of { "sort": { "_shard_doc": "asc" } } will be added to the query. By default, Elasticsearch adds a sort field "_shard_doc" on top of existing sort implicitly as a tie-breakers.

The scroll search has been refactored and is expected to function the same, with only minor changes made to the logging messages.

Co-authored-by: João Duarte <jsvd@users.noreply.github.com>