Support aggregations (#202)
Add `response_type` configuration option to allow processing result of aggregations.

The default `hits` will generate one event per returned document (i.e. "hit"), which is the current behavior.

When set to `aggregations`, a single Logstash event will be generated with the contents of the `aggregations` object of the query's response. In this case the `hits` object will be ignored.
The parameter `size` will be always be set to 0 regardless of the default or user-defined value set in this plugin.

---------
Co-authored-by: MikhailMS <1488maiklm@gmail.com>